Moved get_random_graph to common utils
